+++ android/src/main/java/com/reactnativestripesdk/StripeSdkModule.kt
@@ -13,7 +13,6 @@ import androidx.core.net.toUri
 import androidx.fragment.app.FragmentActivity
 import com.facebook.react.ReactActivity
 import com.facebook.react.bridge.Arguments
-import com.facebook.react.bridge.BaseActivityEventListener
 import com.facebook.react.bridge.Promise
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.bridge.ReactMethod
@@ -98,9 +97,6 @@ class StripeSdkModule(
   private var stripeAccountId: String? = null
   private var urlScheme: String? = null
 
-  private var createPlatformPayPaymentMethodPromise: Promise? = null
-  private var platformPayUsesDeprecatedTokenFlow = false
-
   private val stripeUIManagers = mutableListOf<StripeUIManager>()
   private var paymentSheetManager: PaymentSheetManager? = null
   private var paymentLauncherManager: PaymentLauncherManager? = null
@@ -124,37 +120,6 @@ class StripeSdkModule(
   private val pendingUrlsLock = Any()
 
   val eventEmitter: EventEmitterCompat by lazy { EventEmitterCompat(reactApplicationContext) }
-
-  private val mActivityEventListener =
-    object : BaseActivityEventListener() {
-      override fun onActivityResult(
-        activity: Activity,
-        requestCode: Int,
-        resultCode: Int,
-        data: Intent?,
-      ) {
-        if (::stripe.isInitialized) {
-          when (requestCode) {
-            GooglePayRequestHelper.LOAD_PAYMENT_DATA_REQUEST_CODE -> {
-              createPlatformPayPaymentMethodPromise?.let {
-                GooglePayRequestHelper.handleGooglePaymentMethodResult(
-                  resultCode,
-                  data,
-                  stripe,
-                  platformPayUsesDeprecatedTokenFlow,
-                  it,
-                )
-                createPlatformPayPaymentMethodPromise = null
-              }
-            }
-          }
-        }
-      }
-    }
-
-  init {
-    reactContext.addActivityEventListener(mActivityEventListener)
-  }
 
   override fun invalidate() {
     super.invalidate()
@@ -918,16 +883,15 @@ class StripeSdkModule(
         )
         return
       }
-    platformPayUsesDeprecatedTokenFlow = usesDeprecatedTokenFlow
-    createPlatformPayPaymentMethodPromise = promise
-    getCurrentActivityOrResolveWithError(promise)?.let {
-      val request =
-        GooglePayRequestHelper.createPaymentRequest(
-          it,
+    UiThreadUtil.runOnUiThread {
+      getCurrentActivityOrResolveWithError(promise)?.let { activity ->
+        val request = GooglePayRequestHelper.createPaymentRequest(
+          activity,
           GooglePayJsonFactory(reactApplicationContext),
           googlePayParams,
         )
-      GooglePayRequestHelper.createPaymentMethod(request, it)
+        GooglePayRequestHelper.createPaymentMethod(request, activity, stripe, usesDeprecatedTokenFlow, promise)
+      }
     }
   }